# Conflicts:
#	src/main/java/com/example/capstoneproject/controller/CvController.java

+++ src/main/java/com/example/capstoneproject/service/impl/ProjectServiceImpl.java
@@ -243,7 +243,12 @@ public class ProjectServiceImpl extends AbstractBaseService<Project, ProjectDto,
                 scoreRepository.deleteScoreById(score.getId());
             }
 
-            sectionLogRepository.deleteBySection_Id(section.getId());
+            if(section!=null){
+                sectionLogRepository.deleteBySection_Id(section.getId());
+                cv.setOverview(null);
+                cvRepository.save(cv);
+            }
+
             //Get process evaluate
             List<BulletPointDto> evaluateResult = evaluateService.checkSentences(dto.getDescription());
             ProjectViewDto projectViewDto = new ProjectViewDto();
@@ -263,7 +268,7 @@ public class ProjectServiceImpl extends AbstractBaseService<Project, ProjectDto,
                 Evaluate evaluate = evaluates.get(i);
                 BulletPointDto bulletPointDto = evaluateResult.get(i);
                 SectionLogDto sectionLogDto1 = new SectionLogDto();
-                sectionLogDto1.setSection(sectionMapper.mapDtoToEntity(sectionMapper.mapEntityToDto(section)));
+                sectionLogDto1.setSection(section);
                 sectionLogDto1.setEvaluate(evaluate);
                 sectionLogDto1.setBullet(bulletPointDto.getResult());
                 sectionLogDto1.setStatus(bulletPointDto.getStatus());
